Refactoring code - separating interfaces from services
