+++ CycloneDX.Tests/DotnetUtilsServiceTests.cs
@@ -15,12 +15,9 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (c) OWASP Foundation. All Rights Reserved.
 
-using System;
-using System.Collections.Generic;
-using System.IO.Abstractions;
 using System.IO.Abstractions.TestingHelpers;
+using CycloneDX.Interfaces;
 using XFS = System.IO.Abstractions.TestingHelpers.MockUnixSupport;
-using System.Threading.Tasks;
 using Xunit;
 using Moq;
 using CycloneDX.Models;
