+++ CycloneDX.Tests/GithubServiceTests.cs
@@ -15,7 +15,6 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (c) OWASP Foundation. All Rights Reserved.
 
-using System;
 using System.Threading.Tasks;
 using Xunit;
 using RichardSzalay.MockHttp;
