+++ CycloneDX.Tests/ProgramTests.cs
@@ -22,8 +22,8 @@ using System.Threading.Tasks;
 using XFS = System.IO.Abstractions.TestingHelpers.MockUnixSupport;
 using Moq;
 using CycloneDX.Models;
-using CycloneDX.Services;
 using System.IO;
+using CycloneDX.Interfaces;
 using Microsoft.DotNet.PlatformAbstractions;
 
 namespace CycloneDX.Tests
