+++ CycloneDX.Tests/SolutionFileServiceTests.cs
@@ -15,11 +15,11 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (c) OWASP Foundation. All Rights Reserved.
 
-using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Xunit;
 using System.IO.Abstractions.TestingHelpers;
+using CycloneDX.Interfaces;
 using XFS = System.IO.Abstractions.TestingHelpers.MockUnixSupport;
 using Moq;
 using CycloneDX.Services;
