+++ CycloneDX/Interfaces/IDotnetCommandService.cs
@@ -17,7 +17,7 @@
 
 using CycloneDX.Models;
 
-namespace CycloneDX.Services
+namespace CycloneDX.Interfaces
 {
     public interface IDotnetCommandService
     {
