+++ CycloneDX/Interfaces/INugetService.cs
@@ -1,0 +1,11 @@
+using System.Threading.Tasks;
+using CycloneDX.Models;
+
+namespace CycloneDX.Interfaces
+{
+    public interface INugetService
+    {
+        Task<Component> GetComponentAsync(string name, string version, Component.ComponentScope? scope);
+        Task<Component> GetComponentAsync(NugetPackage nugetPackage);
+    }
+}
