+++ CycloneDX/Services/DotnetCommandService.cs
@@ -20,6 +20,7 @@ using System.Diagnostics.Contracts;
 using System.IO;
 using System.Text;
 using System.Threading.Tasks;
+using CycloneDX.Interfaces;
 using McMaster.Extensions.CommandLineUtils;
 using CycloneDX.Models;
