+++ CycloneDX/Services/ProjectAssetsFileService.cs
@@ -15,9 +15,9 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (c) OWASP Foundation. All Rights Reserved.
 
-using System;
 using System.Collections.Generic;
 using System.IO.Abstractions;
+using CycloneDX.Interfaces;
 using AssetFileReader = NuGet.ProjectModel.LockFileFormat;
 using CycloneDX.Models;
